// src/api/models/create-limit-order-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * IMX Market Trader
 * IMX Market Trader backend API
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface CreateLimitOrderDto
 */
export interface CreateLimitOrderDto {
    /**
     * 
     * @type {string}
     * @memberof CreateLimitOrderDto
     */
    walletId: string;
    /**
     * 
     * @type {string}
     * @memberof CreateLimitOrderDto
     */
    walletAddress?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateLimitOrderDto
     */
    collectionId: string;
    /**
     * 
     * @type {string}
     * @memberof CreateLimitOrderDto
     */
    collectionAddress?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateLimitOrderDto
     */
    assetName: string;
    /**
     * 
     * @type {string}
     * @memberof CreateLimitOrderDto
     */
    orderType: CreateLimitOrderDtoOrderTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof CreateLimitOrderDto
     */
    priceLevel: string;
    /**
     * 
     * @type {number}
     * @memberof CreateLimitOrderDto
     */
    priceLevelMultiplier?: number;
    /**
     * 
     * @type {number}
     * @memberof CreateLimitOrderDto
     */
    feesPercent: number;
    /**
     * 
     * @type {number}
     * @memberof CreateLimitOrderDto
     */
    maxQuantity: number;
}

/**
    * @export
    * @enum {string}
    */
export enum CreateLimitOrderDtoOrderTypeEnum {
    Buy = 'Buy',
    Sell = 'Sell'
}