// src/api/models/limit-order.ts
/* tslint:disable */
/* eslint-disable */
/**
 * IMX Market Trader
 * IMX Market Trader backend API
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface LimitOrder
 */
export interface LimitOrder {
    /**
     * 
     * @type {string}
     * @memberof LimitOrder
     */
    _id: string;
    /**
     * 
     * @type {string}
     * @memberof LimitOrder
     */
    walletId: string;
    /**
     * 
     * @type {string}
     * @memberof LimitOrder
     */
    walletAddress: string;
    /**
     * 
     * @type {string}
     * @memberof LimitOrder
     */
    collectionId: string;
    /**
     * 
     * @type {string}
     * @memberof LimitOrder
     */
    collectionAddress: string;
    /**
     * 
     * @type {string}
     * @memberof LimitOrder
     */
    assetName: string;
    /**
     * 
     * @type {string}
     * @memberof LimitOrder
     */
    orderType: LimitOrderOrderTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof LimitOrder
     */
    priceLevel: string;
    /**
     * 
     * @type {number}
     * @memberof LimitOrder
     */
    priceLevelMultiplier: number;
    /**
     * 
     * @type {number}
     * @memberof LimitOrder
     */
    feesPercent: number;
    /**
     * 
     * @type {number}
     * @memberof LimitOrder
     */
    maxQuantity: number;
}

/**
    * @export
    * @enum {string}
    */
export enum LimitOrderOrderTypeEnum {
    Buy = 'Buy',
    Sell = 'Sell'
}